Move admin into peachjam
